@@ -17,9 +17,8 @@
     "@types/node": "18.13.0",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
-    "dayjs": "^1.11.7",
     "eslint": "8.34.0",
-    "next": "13.2.0",
+    "next": "13.4.3",
     "next-mdx-remote": "^4.3.0",
     "plaiceholder": "^2.5.0",
     "react": "18.2.0",
@@ -41,13 +40,13 @@
     "@typescript-eslint/parser": "^5.52.0",
     "@vitejs/plugin-react": "^3.1.0",
     "autoprefixer": "^10.4.13",
-    "eslint-config-next": "13.2.0",
+    "eslint-config-next": "13.4.3",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-react-hooks": "^4.6.0",
     "husky": "^8.0.3",
     "jsdom": "^21.1.0",
-    "lint-staged": "^13.1.2",
+    "lint-staged": "^13.2.0",
     "next-sitemap": "^3.1.49",
     "postcss": "^8.4.21",
     "prettier": "2.8.4",
