@@ -10,10 +10,13 @@ type ButtonVariants =
   | "tag"
   | "share-social"
   | "dark"
-  | "dark-outline";
+  | "dark-outline"
+  | "footer-link"
+  | "social-menu-link"
+  | "social-menu-link-mobile-nav";
 
 const baseButtonStyles =
-  "inline-flex transition-colors border-2 rounded px-4 md:px-6 py-2 md:py-3 text-base text-center disabled:cursor-not-allowed rounded flex flex-row items-center justify-center gap-1.5";
+  "inline-flex transition-colors border-2 rounded px-4 md:px-6 py-2 md:py-3 text-base text-center disabled:cursor-not-allowed rounded items-center justify-center gap-1.5";
 
 const buttonVariants = {
   white: twMerge(
@@ -43,12 +46,15 @@ const buttonVariants = {
   tag: "inline-flex transition-colors rounded px-2.5 py-1.5 text-xs bg-zinc-200 dark:bg-zinc-700 hover:bg-green-500 dark:hover:bg-green-500 text-zinc-500 dark:text-zinc-400 hover:text-white dark:hover:text-zinc-800",
   "share-social":
     "inline-flex transition-colors border-2 rounded px-3 md:px-4 py-2 md:py-3 bg-green-500 hover:bg-green-600 text-white border-green-500 ",
+  "footer-link": "hover:underline dark:text-zinc-200",
+  "social-menu-link": "text-green-500 hover:text-green-700 transition-colors",
+  "social-menu-link-mobile-nav":
+    "text-white dark:text-zinc-800 hover:text-green-500 dark:hover:text-white lg:text-zinc-600 md:hover:text-green-500 lg:dark:text-white lg:dark:hover:text-green-500 transition-colors",
 } as const;
 
 export type LinkProps = {
   href: string;
   children: ReactNode;
-  className?: string;
   tabIndex?: number;
   onClick?: () => void;
   buttonStyle?: ButtonVariants;
@@ -66,7 +72,6 @@ export const Link = (props: LinkProps) => {
       onClick={props.onClick}
       className={twMerge(
         props.buttonStyle && buttonVariants[props.buttonStyle],
-        props.className,
       )}
       tabIndex={props.tabIndex}
       aria-hidden={props["aria-hidden"]}
@@ -82,7 +87,6 @@ export const Link = (props: LinkProps) => {
       onClick={props.onClick}
       className={twMerge(
         props.buttonStyle && buttonVariants[props.buttonStyle],
-        props.className,
       )}
     >
       {props.children}
