@@ -1,5 +1,6 @@
-import { Container } from "@/components/atoms/Container/Container";
 import { ReactNode } from "react";
+import { getSlug } from "@/utils/getSlug";
+import { Section } from "@/components/molecules/Section/Section";
 
 type BannerProps = {
   title: string;
@@ -9,18 +10,25 @@ type BannerProps = {
 
 export const Banner = ({ title, content, buttons }: BannerProps) => {
   return (
-    <section className="py-24 bg-green-500">
-      <Container>
-        <header className="w-full md:w-2/3 text-center mx-auto text-white flex flex-col gap-5">
-          <h2 className="text-3xl font-bold dark:text-zinc-800">{title}</h2>
-          {content && <p className="dark:text-zinc-800">{content}</p>}
-          {buttons && (
-            <div className="flex flex-row flex-wrap gap-2 justify-center items-center">
-              {buttons}
-            </div>
-          )}
-        </header>
-      </Container>
-    </section>
+    <Section
+      ariaLabelledBy={getSlug(title)}
+      bgColor={{ light: "green", dark: "green" }}
+      paddingVariant="fixed"
+    >
+      <header className="w-full md:w-2/3 text-center mx-auto text-white flex flex-col gap-5">
+        <h2
+          className="text-3xl font-bold dark:text-zinc-800"
+          id={getSlug(title)}
+        >
+          {title}
+        </h2>
+        {content && <p className="dark:text-zinc-800">{content}</p>}
+        {buttons && (
+          <div className="flex flex-row flex-wrap gap-2 justify-center items-center">
+            {buttons}
+          </div>
+        )}
+      </header>
+    </Section>
   );
 };
