@@ -1,4 +1,3 @@
-import { Link } from "@/components/atoms/Link/Link";
 import { getSlug } from "@/utils/getSlug";
 
 type HeadingProps = {
@@ -16,14 +15,15 @@ export const Heading = ({
     const slug = `${getSlug(children)}`;
     return (
       <Tag {...props} id={slug} className="flex gap-3 relative ml-5 lg:ml-0">
-        <Link
+        {/* @todo replace A tag with next/lint or my custom Link component what uses next/link when https://github.com/vercel/next.js/issues/44295 issue will be resolved */}
+        <a
           href={`#${slug}`}
           className="absolute left-0 top-0 -translate-x-full pr-2 transition-all text-neutral-300 dark:text-neutral-600 hover:text-green-500 dark:hover:text-green-500"
           aria-hidden="true"
           tabIndex={-1}
         >
           #
-        </Link>
+        </a>
         {children}
       </Tag>
     );
