@@ -5,12 +5,7 @@ import { z } from "zod";
 import { Post } from "@/types";
 import { getPlaiceholder } from "plaiceholder";
 import { getSlug } from "@/utils/getSlug";
-
-import dayjs from "dayjs";
-import customParseFormat from "dayjs/plugin/customParseFormat";
-dayjs.extend(customParseFormat);
-import "dayjs/locale/pl";
-dayjs.locale("pl");
+import { getISODateFromPublicatedDate } from "@/utils/getISODateFromPublicationDate";
 
 const postsDir = path.join(process.cwd(), "src/content/posts");
 
@@ -75,10 +70,10 @@ export const getAllPosts = async () => {
     slugs.map(async (slug) => await getPostBySlug(slug)),
   );
   const postSortedByDate = posts.sort((a, b) => {
-    const first = dayjs(a.frontmatter.date, "YYYY.MM.DD");
-    const second = dayjs(b.frontmatter.date, "YYYY.MM.DD");
+    const first = getISODateFromPublicatedDate(a.frontmatter.date);
+    const second = getISODateFromPublicatedDate(b.frontmatter.date);
 
-    return first.isBefore(second) ? 1 : -1;
+    return first < second ? 1 : -1;
   });
 
   return postSortedByDate;
@@ -90,11 +85,9 @@ export const getPublishedPosts = async () => {
   return posts.filter((post) => post.frontmatter.published === true);
 };
 
-export const getCategories = async () => {
+export const getPostsParams = async () => {
   const posts = await getPublishedPosts();
-  const categories = new Set(posts.map((post) => post.frontmatter.category));
-
-  return Array.from(categories);
+  return posts.map((post) => ({ slug: post.frontmatter.slug }));
 };
 
 export const getTags = async () => {
@@ -104,18 +97,56 @@ export const getTags = async () => {
   return Array.from(tags);
 };
 
+export const getTagsParams = async () => {
+  const tags = await getTags();
+
+  return tags.map((tag) => ({ slug: getSlug(tag) }));
+};
+
+export const getTagBySlug = async (slug: string) => {
+  const tags = await getTags();
+  const fullTagName = tags.find((tag) => getSlug(tag) === slug);
+
+  return fullTagName ? fullTagName : slug;
+};
+
 export const getPostsByTag = async (givenTag: string) => {
   const posts = await getPublishedPosts();
-  return posts.filter((post) =>
-    post.frontmatter.tags.map((tag) => getSlug(tag)).includes(givenTag),
-  );
+  return {
+    posts: posts.filter((post) =>
+      post.frontmatter.tags.map((tag) => getSlug(tag)).includes(givenTag),
+    ),
+    tag: await getTagBySlug(givenTag),
+  };
+};
+export const getCategories = async () => {
+  const posts = await getPublishedPosts();
+  const categories = new Set(posts.map((post) => post.frontmatter.category));
+
+  return Array.from(categories);
+};
+
+export const getCategoriesParams = async () => {
+  const categories = await getCategories();
+
+  return categories.map((cat) => ({ slug: getSlug(cat) }));
+};
+
+export const getCategoryBySlug = async (slug: string) => {
+  const categories = await getCategories();
+  const fullCategoryName = categories.find((cat) => getSlug(cat) === slug);
+
+  return fullCategoryName ? fullCategoryName : slug;
 };
 
 export const getPostsByCategory = async (givenCategory: string) => {
   const posts = await getPublishedPosts();
-  return posts.filter(
-    (post) => getSlug(post.frontmatter.category) == givenCategory,
-  );
+  return {
+    posts: posts.filter(
+      (post) => getSlug(post.frontmatter.category) == givenCategory,
+    ),
+    category: await getCategoryBySlug(givenCategory),
+  };
 };
 
 export const getRelatedPostsByCategory = async (
